+++ Assets/Scripts/Weapon/Weapon/Gun.cs
@@ -1,0 +1,55 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Gun : Weapon
+{
+    [SerializeField]
+    private SpriteRenderer spriteRenderer;
+
+    [SerializeField]
+    private GameObject bulletPrefab;  // ÃÑ¾Ë ÇÁ¸®ÆÕ
+
+    [SerializeField]
+    private Transform ShootStartPoint;
+
+    private ProjectilePoolManager projectilePoolManager;
+
+    void Start()
+    {
+        projectilePoolManager = ProjectilePoolManager.Instance;
+    }
+
+    void Update()
+    {
+        RotateWeaponTowardsMouse();
+    }
+
+    protected override void PerformAttack()
+    {
+        projectilePoolManager.GetProjectileFromPool(bulletPrefab, ShootStartPoint.position, transform.rotation);
+    }
+
+
+    void RotateWeaponTowardsMouse()
+    {
+        Vector3 mousePosition = Camera.main.ScreenToWorldPoint(Input.mousePosition);
+        mousePosition.z = 0f;
+
+        Vector3 weaponPosition = transform.position;
+        weaponPosition.z = 0f;
+
+        Vector3 direction = mousePosition - weaponPosition;
+        float angle = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
+
+        if (direction.x < 0f) {
+            spriteRenderer.flipY = true;
+
+        }
+        else {
+            spriteRenderer.flipY = false;
+
+        }
+        transform.rotation = Quaternion.AngleAxis(angle, Vector3.forward);
+    }
+}
