+++ Assets/Scripts/Weapon/Weapon/Gun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7944d956ea8dc72459327d29488fa86e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
