+++ Assets/Scripts/Weapon/Weapon/MeleeWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df2e30db36e2bde458590bae996e8d6c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
